+++ fed-challenge/code-breaker/src/index.js
@@ -1,21 +1,17 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { BrowserRouter as Router } from "react-router-dom";
-
 import { Provider } from "react-redux";
 import store from "./store";
+import App from "./App";
+import { GlobalStyles } from "./Shared.styles";
 
 import "bootstrap/dist/css/bootstrap.min.css";
-import "./index.css";
-
-import App from "./App";
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
-      <Router>
-        <App />
-      </Router>
+      <GlobalStyles />
+      <App />
     </Provider>
   </React.StrictMode>,
   document.getElementById("root")
